\name{base.stacking.thermodynamics}
\alias{base.stacking.thermodynamics}
\docType{data}
\title{ Base-stacking thermodynamic parameters}
\description{
  Unified thermodynamic parameters (delta-S and delta-H values) for nearest-neighbor base stacking calculations
}
\usage{data(base.stacking.thermodynamics)}
\format{
  A matrix with 2 columns and 16 rows.  Column 1 indicates enthalpic parameters (dH) and column 2 indicates entropic parameters (dS).  Rows indicate all possible 2bp combinations of "A", "T", "C", and "G" (e.g. "AC")
}
\source{
  Allawi, H.T. and SantaLucia, J., Jr. (1997) Thermodynamics and NMR of internal G.T mismatches in DNA, \emph{Biochemistry}, \strong{36}, 10581-10594.
}
\examples{
data(base.stacking.thermodynamics)
}
\keyword{datasets}
