\name{calcGC-methods}
\docType{methods}
\alias{calcGC-methods}
\alias{calcGC,missing-method}
\alias{calcGC,NULL-method}
\alias{calcGC,character-method}
\alias{calcGC,ExpressionSet-method}
\title{ Calculate GC percent (methods)}
\description{
 Methods for calculating GC percent from oligonucleotide sequences
}
\section{Methods}{
\describe{

\item{x = "missing"}{ Handle empty function call }

\item{x = "NULL"}{ Handle empty function call }

\item{x = "character"}{ Handle character input }

\item{x = "ExpressionSet"}{ Handle input of an object of class \code{ExpressionSet} }
}}
\seealso{ \code{\link{calcGC}} }
\keyword{methods}
