\name{calcTm-methods}
\docType{methods}
\alias{calcTm-methods}
\alias{calcTm,missing-method}
\alias{calcTm,NULL-method}
\alias{calcTm,character-method}
\alias{calcTm,ExpressionSet-method}
\title{ Calculate Tm (methods)}
\description{
 Methods for calculating melting temperature (Tm) of nucleotide sequences
}
\section{Methods}{
\describe{

\item{x = "missing"}{ Handle empty function call }

\item{x = "NULL"}{ Handle empty function call }

\item{x = "character"}{ Handle character input }

\item{x = "ExpressionSet"}{ Handle input of an object of class \code{ExpressionSet} }
}}
\seealso{ \code{\link{calcTm}}}
\keyword{methods}
