\name{exprs2-methods}
\docType{methods}
\alias{exprs2-methods}
\alias{exprs2,missing-method}
\alias{exprs2,ExpressionSet-method}
\alias{exprs2<-,ExpressionSet,missing-method}
\alias{exprs2<-,ExpressionSet,matrix-method}
\title{ Retrieve microarray data (for signal channel 2) from ExpressionSets (methods)}
\description{
 Methods for accessing (and/or assigning) data for signal channel 2 in a given ExpressionSet object
}
\section{Methods}{
\describe{

\item{object = "missing"}{ Handle empty function call }

\item{object = "ExpressionSet"}{ Handle input of an object of class \code{ExpressionSet} }

\item{object = "ExpressionSet", value = "missing"}{ Handle empty function call }

\item{object = "ExpressionSet", value = "matrix"}{ Handle input of an object of class \code{ExpressionSet}e }
}}
\seealso{ \code{\link{exprs2}} }
\keyword{methods}
