% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geology.R
\name{get_geology}
\alias{get_geology}
\title{Get the geology of South Tyrol for an object of class sf}
\usage{
get_geology(
  sf.object = NULL,
  geo_path = "/mnt/CEPH_PROJECTS/Proslide/Envir_data/Geology/reclass_geo.gpkg"
)
}
\description{
Get the geology of South Tyrol for an object of class sf
}
