% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savegg.R
\name{savegg}
\alias{savegg}
\title{Save ggplot to multiple devices}
\usage{
savegg(f = NULL, dev = NULL, ...)
}
\arguments{
\item{f}{Filename without any extension}

\item{dev}{Vector of entensions}
}
\description{
Save ggplot to multiple devices
}
