% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{seasons}
\alias{seasons}
\title{Return northern hemisphere season for a date}
\usage{
seasons(date)
}
\arguments{
\item{date}{An object of class \code{Date}, \code{POSIXct} or an integer beteween 1 and 12}
}
\description{
Return northern hemisphere season for a date
}
