% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ma.R
\name{theme_ma}
\alias{theme_ma}
\title{The ggplot theme for my Masters Thesis}
\usage{
theme_ma(
  base_size = 13,
  base_family = "Times",
  title_size = 18,
  subtitle_size = 13
)
}
\description{
The ggplot theme for my Masters Thesis
}
