% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr.R
\name{wr}
\alias{wr}
\title{Write or append to file}
\usage{
wr(file, text, mode = NULL)
}
\arguments{
\item{file}{The path to the existent or non-existent file}

\item{text}{The text to either append or write}

\item{mode}{Append (a) or write (w)}
}
\description{
Write or append to file
}
